+++ app/src/main/java/com/example/common_project01/MainActivity.kt
@@ -25,7 +25,7 @@ class MainActivity : AppCompatActivity() {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(setOf(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications))
+                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_friends))
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
