+++ app/src/main/java/com/example/common_project01/ui/friends/Friend.kt
@@ -1,0 +1,3 @@
+package com.example.common_project01.ui.friends
+
+data class Friend(val image: Int, val name: String, val id:String, val intro: String, val profile: Boolean)
