--- app/src/main/java/com/example/common_project01/ui/notifications/Friend.kt
@@ -1,3 +1,0 @@
-package com.example.common_project01.ui.notifications
-
-data class Friend(val image: Int, val name: String, val id:String, val intro: String)
