+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -12,7 +12,7 @@
         android:title="@string/title_dashboard" />
 
     <item
-        android:id="@+id/navigation_notifications"
+        android:id="@+id/navigation_friends"
         android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/title_notifications" />
