+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -18,8 +18,8 @@
         tools:layout="@layout/fragment_dashboard" />
 
     <fragment
-        android:id="@+id/navigation_notifications"
-        android:name="com.example.common_project01.ui.notifications.FriendListFragment"
+        android:id="@+id/navigation_friends"
+        android:name="com.example.common_project01.ui.friends.FriendListFragment"
         android:label="@string/title_notifications"
         tools:layout="@layout/fragment_friend_list" />
 </navigation>
