+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,6 @@
     <string name="title_home">Home</string>
     <string name="title_dashboard">Dashboard</string>
     <string name="title_notifications">Notifications</string>
+    <string name="profile_text">프로필</string>
+    <string name="friend_text">친구 목록</string>
 </resources>
